After implementing question, one percent brokerage and row chart

+++ otree_single_asset_market/static/otree_single_asset_market/order_enter_widget.js
@@ -99,7 +99,9 @@ class OrderEnterWidget extends PolymerElement {
     }
 
     _enter_order(event) {
-        const price = parseInt(this.$.price_input.value);
+        // const price = parseInt(this.$.price_input.value);
+        // Changed by Kaushik to implement float price
+        const price = parseFloat(this.$.price_input.value);
         const volume = parseInt(this.$.volume_input.value);
         const is_bid = (event.target.value == "bid");
         const order = {
